feat(console): init connector page
