+++ packages/console/src/App.tsx
@@ -11,6 +11,7 @@ import Topbar from './components/Topbar';
 import initI18n from './i18n/init';
 import ApiResources from './pages/ApiResources';
 import Applications from './pages/Applications';
+import Connector from './pages/Connectors/Connector';
 import { fetcher } from './swr';
 
 const isBasenameNeeded = process.env.NODE_ENV !== 'development' || process.env.PORT === '5002';
@@ -37,6 +38,7 @@ const Main = () => {
             <Routes>
               <Route path="api-resources" element={<ApiResources />} />
               <Route path="applications" element={<Applications />} />
+              <Route path="connectors/:connectorId" element={<Connector />} />
             </Routes>
           </Content>
         </div>
