+++ packages/schemas/src/types/index.ts
@@ -1,1 +1,2 @@
 export * from './user';
+export * from './connector';
